setu-backend/requirements.txt
# Rural AI Assistant Backend — production dependencies

# Framework
fastapi>=0.109.0,<1.0
uvicorn[standard]>=0.27.0,<1.0

# Database
sqlalchemy>=2.0.0,<3.0
psycopg2-binary>=2.9.9,<3.0

# AWS
boto3>=1.34.0,<2.0

# Config & validation
python-dotenv>=1.0.0,<2.0
pydantic>=2.5.0,<3.0
pydantic-settings>=2.1.0,<3.0

# Auth
bcrypt>=4.0.0,<5.0
python-jose[cryptography]>=3.3.0

# Voice: Whisper (free STT), Polly via boto3 (TTS)
faster-whisper>=1.0.0,<2.0

# Optional: Google STT (when STT_PROVIDER=google, requires billing)
# google-cloud-speech>=2.24.0,<3.0

# Twilio WhatsApp
twilio>=8.0.0,<10.0

# Optional for future use (uncomment when needed):
# langgraph
# pinecone-client
